Change user_flash

+++ 2.Firmware/1.FrontCar/MM32F527X_E9P/project/code/easy_ui.h
@@ -72,7 +72,7 @@ extern uint8_t opnEnter, opnExit, opnUp, opnDown;
 #ifdef FPU
 typedef     double      paramType;
 #else
-typedef     int32_t     paramType;
+typedef     int64_t     paramType;
 #endif
 
 typedef enum
